requirements.txt
flask
dnspython